requirements.txt
# Core Framework
Flask==3.0.0
Flask-CORS==4.0.0

# Database
supabase>=2.3.0

# AI & ML (lighter versions for Vercel)
google-generativeai>=0.3.0
sentence-transformers>=2.2.2
numpy>=1.24.0

# Utilities
python-dotenv>=1.0.0
